agents/01-02-langchain-start/requirements.txt
langchain
langchain-openai
langchain-google-genai
openai
google-generativeai
python-dotenv